This is just breaking the existing MinecraftOres.xml up into parts. There should be zero changes to the world.
The main config file has been changed from explicitly listing the files to include
and now uses a *.xml import instead.

Again: There should be no changes to any generation by this.
This is not the final breakup/split, just the first version.
See the README files for more information.