synchronize generation to hopefully fix #171
Generation of a distribution is not the same as population. While
population should only happen during world generation, distributions
need to be generated in order to be rendered in debugging mode. This
means that it is convenient and probably sensible to generate
distributions in the server network handling thread, upon receipt of a
geometry request.